Add threshold value in GUI for framefinder threshold

+++ src/CellSorter/lib/framefinder.h
@@ -43,7 +43,7 @@ int get_files(std::vector<Frame>& files, const std::string& folder);
 void accept_or_reject(std::vector<Frame>& frames, const std::string& img_folder,
                       const double& threshold);
 
-bool hasChanged(const cv::Mat& img1, const cv::Mat& img2, const double& threshold);
+bool hasChanged(const cv::Mat& img1, const cv::Mat& img2, const int& threshold);
 void get_accepted(const std::vector<Frame>& frames, std::vector<Frame>& output);
 void get_rejected(const std::vector<Frame>& frames, std::vector<Frame>& output);
